refactor(ui): co-locate the skills block with the resource drawer
